product-search-python/backup/requirements.txt
flask
gunicorn
torch
torchvision
ftfy
regex
tqdm
git+https://github.com/openai/CLIP.git
pillow
requests